exercise-bt-collections-arraylistsort/src/main/java/com/stackroute/collections/ArrayListSort.java
package com.stackroute.collections;

import java.math.BigInteger;
import java.util.ArrayList;
import java.util.Collections;
import java.util.Comparator;

public class ArrayListSort {
    //write logic to sort given list in descending order
    public ArrayList<BigInteger> arrayListSortDesc(ArrayList<BigInteger> list) {

        Comparator<BigInteger> result = new Comparator<BigInteger>() {
            @Override
            public int compare(BigInteger o1, BigInteger o2) {
                return o2.compareTo(o1);
            }
        };

        list.sort(result);
        return list;

    }
}
